control/pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Hexitec-MHz-Top-Level-Control"
description = "odin-control based system for managing aqusitions and DAQ system state for the Hexitec-MHz Detector system"
dynamic = ["version"]
requires-python = ">=3.10.0"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "odin-control @ git+https://github.com/odin-detector/odin-control@1.6.0#egg=odin-control",
    "requests",
    "python-statemachine"
]
readme = "README.md"

[[project.authors]] 
email = "josh.harris@stfc.ac.uk"
name = "Josh Harris"

[project.urls]
GitHub = "https://github.com/stfc-aeg/hexitec-mhz-detector"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
# Optional settings for setuptools_scm
root = ".."
write_to = "control/src/hexitec/_version.py"
version_scheme = "no-guess-dev"
local_scheme = "node-and-date"






